Update hasNecessarySupervisionComponent to include the MATCH_DISABLED_COMPONENTS flag.
For some reason, in practice, MATCH_ALL doesn't seem to include all the
effects of MATCH_DISABLED_COMPONENTS and does not return a result when
the package is disabled. Unfortunately, the unit test also doesn't seem
to be able to catch this case.

Bug: 422632752
Test: atest SupervisionHelperTest
Flag: android.app.supervision.flags.enable_supervision_settings_screen
Change-Id: I1049b1684ff0a898ab92f9ed7e1447c17affa9b7