Fix missing views on display listeners update
onAttach might be called before/after onBindViewHolder, and hence
this should be guarded properly.

This hasn't happened during development because when this race condition
happened, topology is never null and it went to displayBlocks() which
already has a guard for paneContent.isInitialized. However, when
topology is null, it updates topologyHint and the crash happened.

Bug: 427858975
Test: atest DisplayTopologyPreferenceControllerTest
Flag: EXEMPT bugfix
Change-Id: Ib35269fa73ac6aadf81ca2d3c236cd50c6c42340