Update callback invoke condition for battery usage chart.
- Replace isResumed() with isVisible().
This check is to skip UI change caused by callbacks when the user is not keep on the battery usage screen. We used to use Resume status as condition. However, sometimes the callback function maybe invoked with isResumed()=False while switching screen with the swipe back gesture. In this case the usage app list will not updated due to the battery diff map was not saved. So we use `isVisible()` as a more precise condition instead.

Bug: 427900910
Fix: 427900910
Test: manual
Flag: EXEMPT for simple fix
Change-Id: Ia9453caacbb9edc902b3477733a55ed2ed364a48