Add delay for commit changes that would trigger configuration change.
Without the delay, sometime the slider would flicker when clicking the
button consecutively in a short period of time.

From the past in TextReadingPreviewController's history, the flickering
was caused by a race condition: commit a change that could cause
configuration change -> framework takes a snapshot of the screen in
preparation of showing the configuration changes smoother later -> app
updates preview/slider -> fade out the previous taken screenshot and
show the new screen.

The solution is to add a short delay before commit the change, so that
we can be sure that the screenshot accurately reflect the current stage.
Also using the Choregrapher to provide a hint that this delay is needed
to be executed in the next drawing frame.

Bug: 372776688
Test: atest SettingsRoboTests
Test: manually click the button to change the display/font size and make
sure the slider is not flickering.
Flag: EXEMPT not yet used

Change-Id: Ic07fd4504acda9cbc15effee9aeadcc779a8686a