Fix hearing device footer visibility gone issue
AccessibilityFooterPreferenceController will check if title or summary
is empty to update the visibility of the footer preference. However,
hearing device footer's title is set after this check and hence the
visibility is accidentally updated to invisible.

Set the visibility to be visible once we update the title.

Flag: EXEMPT bugfix
Bug: 431669392
Test: manually check
Change-Id: Ib07ae16ad907a019f0a191f35bf81db8007c0f90