Don't clear profile owner on DisableSupervisionActivity
This is actually not possible from within Settings.
It's responsibility of the app to call APIs to clear PO when they wish to.
We'll keep the ability of supervision profile owner to disable supervision, but after that, they should clear PO themselves.

Bug: b/428742276
Flag: android.app.supervision.flags.supervision_manager_apis
Test: atest DisableSupervisionActivityTest
Change-Id: I5126f8b04e077387522f723a7d140261cede8651