Override getValue/setValue when creating custom KeyValueStore
When the datastore is accessed by the app function, it calls the
getValue function. We need to override getValue/setValue instead of
getInt/setInt so that it returns the correct value.

Override the getDefalutValue so that when the value is not set, app
function can still get the default value.

Bug: 372776688
Test: use adb command to trigger app function
Test: use dev tool to modify the setting
Test: atest SettingsRoboTests
Flag: com.android.settings.flags.catalyst_text_reading_screen
Change-Id: Id60113324a79f24ecfea8c70f5b8dec98a5503b5