Use alternative for debounce saving preferred shortcuts
**Root cause**
The passed in coroutineScope may not be lifecycle aware. If we use the
passed in coroutineScope to observe a hot flow, the coroutine would
never ends. This cause the timeout happens when the preference screen is
accessed by app function.

Bug: 434774087
Test: atest SettingsRoboTests
Test: enable the screen in app function and execute the adb command to
make sure it didn't hang
Flag: EXEMPT the code change is behind the caller that are flagged

Change-Id: Ief47e360941a6470c96ee415a0333e1708cbaa9a