Fix missing UI in parental controls dashboard when loading cached preference data.
Fixes a UI flicker on screens that use a cached PreferenceDataProvider. If the cache is empty, it waits for the fresh data before showing anything. If the cache is populated, it shows the cached data immediately and then fetches an update in the background.

Bug: b/433799578, b/433478331
Flag: android.app.supervision.flags.enable_supervision_settings_screen
Test: atest
Change-Id: I3f4a6125ac03ce45248b09d14e18ceafecd0c937