Use removeUserEvenWhenDisallowed when deleting supervising profile
When supervision is disabled, the user restrictions on adding/removing
users are lifted, but when deleting supervision data sometimes this is
not done before we try to delete the supervising profile. Get around
this race condition by removing the user even when disallowed.

Flag: android.app.supervision.flags.enable_supervision_settings_screen
Bug: 436629281
Test: atest SupervisionHelperTest
Change-Id: I10b1f096726dfe696d9f00ca7a5f98fe2191c32f