Hardcode main preference keys in Setup Wizard accessibility fragments.
In SetupWizard, it was using the PreferenceController to get the preference key attached to the screen. With the migration of PreferenceMetadata, the PreferenceController is no longer attached to the screen, hence can't use the PreferenceController to get the key, because the preference controller doesn't exist.

Use the hardcoded preferece key that matches the xml in Setup Wizard.

Bug: 438253464
Test: manual
Test: atest SettingsRoboTests
Flag: EXEMPT low risk bugfix
Change-Id: I9e99b9e0cbc25d4611fe8786349d165f0d286d30