Hide "Default connection behaviour" setting when extended mode is enabled
With the updated external display connection dialog, the dialog is not shown when a device is capable of extended mode. In this case, by default it launches into desktop mode, and they can switch to mirroing using the setting toggle, so we don't want to provide the option to show the dialog using the setting.

Bug: 413620089
Fix: 404330731
Test: manually testing
Test: ExternalDisplayPreferenceFragmentTest
Flag: com.android.window.flags.enable_updated_display_connection_dialog
Change-Id: I86d6d1393a46d871f9c630283ec5f9615c75333a