Observe settings change in onCreate to correctly reflect the state
Root cause:
When the screen is brought back to the foreground, it didn't refetch
the summary data and update the UI. Namely, it didn't execute the
PreferenceMetadata#bind method, hence the summary was not updated.
Moving the observering logic to onCreate, so that it knows data are
changed while the screen is in the background in order to update the
data correctly when the screen is brought to the foreground again.

Bug: 438376003
Test: atest ColorCorrectionScreenTest ColorInversionScreenTest
Test: manual
Flag: EXEMPT BUGFIX
Change-Id: Ic5c7b1425bf8112dd4b1000a969ebddb14983eca