Retain the QS Tooltip regardless on device orientation changes
Root cause:
We need a view (windowToken) in order to show the PopupWindow.
Previously we use the SliderPreference's slider as the view to get the
windowToken to show the PopupWindow. However, there are cases when on
landscape mode, the font size slider may not be initiated yet, since the
font size is not visible yet.

Given we're getting the windowToken in order to show the popup at the
middle top of the screen, not anchoring at the specific view. We could
use the DecorView to get the windowToken and keep the PopupWindow
persists on device orientation change.

Bug: 321173787
Test: manual
Flag: EXEMPT BUGFIX
Change-Id: I1e60c9b0ae08e6563fdf21523250022901138f58